Refactor API tests into separate files by resource.
- Extract createApp function from api.ts for sharing between server and tests
- Split monolithic api.test.ts into resource-specific test files:
  - databases.test.ts for database and connection test endpoints
  - queries.test.ts for query endpoints
  - worksheets.test.ts for worksheet endpoints
  - health.test.ts for health endpoint
- Simplify api-test-helper.ts with setupApiMocks and resetTestDatabase helpers
- Remove mock-adapter.ts (no longer needed)
- Add enableLogging option to createApp for quieter test output
